+++ src/components/BulkEditList/BulkEditListResult/BulkEditListResult.js
@@ -4,7 +4,6 @@ import {
 } from 'react-router-dom';
 
 import { NoResultsMessage } from './NoResultsMessage/NoResultsMessage';
-import { ProgressBar } from '../../shared/ProgressBar/ProgressBar';
 import PreviewContainer from './PreviewContainer/PreviewContainer';
 
 const BulkEditListResult = () => {
@@ -19,10 +18,6 @@ const BulkEditListResult = () => {
         path="/bulk-edit/:id/preview"
         component={PreviewContainer}
         exact
-      />
-      <Route
-        path="/bulk-edit/:id/progress"
-        component={ProgressBar}
       />
     </Switch>
   );